# Conflicts:
#	app/gtnForecasting/src/main/java/com/stpl/app/gtnforecasting/accrualrateprojection/logic/AccrualRateProjectionLogic.java

+++ app/gtnCff/src/main/java/com/stpl/app/cff/queryUtils/CFFQueryUtils.java
@@ -142,7 +142,7 @@ public class CFFQueryUtils {
                 }
             }
         }
-        int count = (Integer) HelperTableLocalServiceUtil.executeUpdateQueryCount(queryString.toString());
+        int count = HelperTableLocalServiceUtil.executeUpdateQueryCount(queryString.toString());
         return count > 0 ? Boolean.TRUE : Boolean.FALSE;
     }
 }
